+++ test/model/tetrimino/type/stetrimino_test.cpp
@@ -58,8 +58,8 @@ TEST_CASE("class STetrimino") {
     REQUIRE(tetrimino.X() == 3);
     REQUIRE(tetrimino.Y() == 0);
 
-    for (int line = 0; line < tetrimino.minos().size(); ++line) {
-      for (int column = 0;
+    for (size_t line = 0; line < tetrimino.minos().size(); ++line) {
+      for (size_t column = 0;
            column < tetrimino.minos().at(tetrimino.orientation())[line].size();
            ++column) {
         REQUIRE(tetrimino.minos().at(tetrimino.orientation())[line][column] ==
@@ -72,8 +72,8 @@ TEST_CASE("class STetrimino") {
       auto tetrimino = STetrimino();
       tetrimino.rotate(true);
 
-      for (int line = 0; line < tetrimino.minos().size(); ++line) {
-        for (int column = 0;
+      for (size_t line = 0; line < tetrimino.minos().size(); ++line) {
+        for (size_t column = 0;
              column <
              tetrimino.minos().at(tetrimino.orientation())[line].size();
              ++column) {
@@ -87,8 +87,8 @@ TEST_CASE("class STetrimino") {
       tetrimino.rotate(true);
       tetrimino.rotate(true);
 
-      for (int line = 0; line < tetrimino.minos().size(); ++line) {
-        for (int column = 0;
+      for (size_t line = 0; line < tetrimino.minos().size(); ++line) {
+        for (size_t column = 0;
              column <
              tetrimino.minos().at(tetrimino.orientation())[line].size();
              ++column) {
@@ -100,8 +100,8 @@ TEST_CASE("class STetrimino") {
     SECTION("check westShape") {
       auto tetrimino = STetrimino();
       tetrimino.rotate(false);
-      for (int line = 0; line < tetrimino.minos().size(); ++line) {
-        for (int column = 0;
+      for (size_t line = 0; line < tetrimino.minos().size(); ++line) {
+        for (size_t column = 0;
              column <
              tetrimino.minos().at(tetrimino.orientation())[line].size();
              ++column) {
